feat: move code to weighting, add tests
