+++ src/semble/ranking/__init__.py
@@ -1,4 +1,5 @@
-from semble.ranking.boosting import apply_query_boost, boost_multi_chunk_files, resolve_alpha
+from semble.ranking.boosting import apply_query_boost, boost_multi_chunk_files
 from semble.ranking.penalties import rerank_topk
+from semble.ranking.weighting import resolve_alpha
 
 __all__ = ["apply_query_boost", "boost_multi_chunk_files", "rerank_topk", "resolve_alpha"]
