+++ tests/test_ranking.py
@@ -1,7 +1,8 @@
 import pytest
 
-from semble.ranking.boosting import apply_query_boost, boost_multi_chunk_files, resolve_alpha
+from semble.ranking.boosting import apply_query_boost, boost_multi_chunk_files
 from semble.ranking.penalties import rerank_topk
+from semble.ranking.weighting import resolve_alpha
 from tests.conftest import make_chunk
 
 
@@ -147,3 +148,9 @@ def test_boost_multi_chunk_files() -> None:
     scores: dict = {c1: 1.0, c2: 0.8, c3: 1.0}
     boost_multi_chunk_files(scores)
     assert scores[c1] > 1.0
+
+
+def test_boosting_with_empty() -> None:
+    """Test that boosting with empty chunks return None."""
+    boosted = apply_query_boost({}, "query", [])
+    assert boosted == {}
